fix: help component should not be rendered by server
